requirements.txt
requests
flask
flask-bootstrap
pyyaml